revert: roll kernel package back twelve hours
