--- src/lingtai_sdk/capabilities/file/__init__.py
@@ -1,1 +1,0 @@
-"""File-I/O tool capabilities: read, write, edit, glob, grep."""
