+++ tests/test_preset_connectivity.py
@@ -178,65 +178,3 @@ def test_check_many_preserves_input_order(monkeypatch):
 def test_check_many_empty_list_returns_empty():
     from lingtai_kernel.preset_connectivity import check_many
     assert check_many([]) == []
-
-
-# ---------------------------------------------------------------------------
-# Local CLI-login providers (e.g. claude-agent-sdk)
-# ---------------------------------------------------------------------------
-
-
-def test_claude_agent_sdk_ok_when_package_importable(monkeypatch):
-    """A local CLI-login provider with its package importable is `ok` —
-    no base_url, no api_key_env, and crucially no TCP probe."""
-    from lingtai_kernel import preset_connectivity
-    with patch.object(preset_connectivity, "_probe_host") as probe, \
-         patch.object(preset_connectivity, "_module_available", return_value=True):
-        result = preset_connectivity.check_connectivity(
-            provider="claude-agent-sdk",
-            base_url=None,
-            api_key_env=None,
-        )
-        assert result["status"] == "ok"
-        probe.assert_not_called()  # local provider — never hits the network
-
-
-def test_claude_agent_sdk_no_base_url_does_not_error(monkeypatch):
-    """The bug: a saved claude-agent-sdk preset was reported unreachable with
-    'no base_url and no default URL'. A local CLI-login provider must never
-    fail just because it has no base_url."""
-    from lingtai_kernel import preset_connectivity
-    with patch.object(preset_connectivity, "_module_available", return_value=True):
-        result = preset_connectivity.check_connectivity(
-            provider="claude-agent-sdk",
-            base_url=None,
-            api_key_env=None,
-        )
-        assert result["status"] != "unreachable"
-        assert "no base_url" not in (result.get("error") or "")
-
-
-def test_claude_agent_sdk_underscore_alias_treated_as_local(monkeypatch):
-    """The underscore alias claude_agent_sdk is the same local provider."""
-    from lingtai_kernel import preset_connectivity
-    with patch.object(preset_connectivity, "_module_available", return_value=True):
-        result = preset_connectivity.check_connectivity(
-            provider="claude_agent_sdk",
-            base_url=None,
-            api_key_env=None,
-        )
-        assert result["status"] == "ok"
-
-
-def test_claude_agent_sdk_missing_package_reports_no_credentials(monkeypatch):
-    """When the optional package is absent, report a clear, actionable status
-    (not the misleading 'no base_url' error)."""
-    from lingtai_kernel import preset_connectivity
-    with patch.object(preset_connectivity, "_module_available", return_value=False):
-        result = preset_connectivity.check_connectivity(
-            provider="claude-agent-sdk",
-            base_url=None,
-            api_key_env=None,
-        )
-        assert result["status"] == "no_credentials"
-        assert "claude-agent-sdk" in (result.get("error") or "")
-        assert "no base_url" not in (result.get("error") or "")
